[update] change the type of register (i32 -> u32)

+++ src/cpu/execution/inst_32.rs
@@ -24,7 +24,7 @@ pub fn exe_inst(inst: &Instruction, cpu: &mut CPU) -> Result<(), (Option<u32>, T
 
     // add the program counter when it isn't a branch instruction
     if cpu.pc == prev_pc {
-        cpu.add2pc(INST_SIZE as i32);
+        cpu.add2pc(INST_SIZE);
     }
 
     Ok(())
