+++ src/cpu/execution/inst_32/base_i.rs
@@ -6,115 +6,115 @@ pub fn exec(inst: &Instruction, cpu: &mut CPU) -> Result<(), (Option<u32>, TrapC
 
     match inst.opc {
         OpecodeKind::OP_LUI => {
-            cpu.regs.write(inst.rd, inst.imm.unwrap());
+            cpu.regs.write(inst.rd, inst.imm.unwrap() as u32);
         },
         OpecodeKind::OP_AUIPC => {
-            cpu.regs.write(inst.rd, cpu.pc as i32 + inst.imm.unwrap());
+            cpu.regs.write(inst.rd, (cpu.pc + inst.imm.unwrap() as u32) as u32);
         },
         OpecodeKind::OP_JAL => {
-            cpu.regs.write(inst.rd, (cpu.pc + INST_SIZE) as i32); 
-            cpu.add2pc(inst.imm.unwrap());
+            cpu.regs.write(inst.rd, cpu.pc + INST_SIZE); 
+            cpu.add2pc(inst.imm.unwrap() as u32);
         },
         OpecodeKind::OP_JALR => {
             // calc next_pc before updated
             let next_pc = cpu.pc + INST_SIZE;
             // setting the least-significant bit of the result to zero-->vvvvvv
-            cpu.update_pc(((cpu.regs.read(inst.rs1) + inst.imm.unwrap()) & !0x1) as u32);
-            cpu.regs.write(inst.rd, next_pc as i32); 
+            cpu.update_pc(((cpu.regs.read(inst.rs1) + inst.imm.unwrap() as u32) & !0x1) as u32);
+            cpu.regs.write(inst.rd, next_pc); 
         },
         OpecodeKind::OP_BEQ => {
             if cpu.regs.read(inst.rs1) == cpu.regs.read(inst.rs2) {
-                cpu.add2pc(inst.imm.unwrap());
+                cpu.add2pc(inst.imm.unwrap() as u32);
             } 
         },
         OpecodeKind::OP_BNE => {
             if cpu.regs.read(inst.rs1) != cpu.regs.read(inst.rs2) {
-                cpu.add2pc(inst.imm.unwrap());
+                cpu.add2pc(inst.imm.unwrap() as u32);
             } 
         },
         OpecodeKind::OP_BLT => {
-            if cpu.regs.read(inst.rs1) < cpu.regs.read(inst.rs2) {
-                cpu.add2pc(inst.imm.unwrap());
+            if (cpu.regs.read(inst.rs1) as i32) < (cpu.regs.read(inst.rs2) as i32) {
+                cpu.add2pc(inst.imm.unwrap() as u32);
             } 
         },
         OpecodeKind::OP_BGE => {
-            if cpu.regs.read(inst.rs1) >= cpu.regs.read(inst.rs2) {
-                cpu.add2pc(inst.imm.unwrap());
+            if (cpu.regs.read(inst.rs1) as i32) >= (cpu.regs.read(inst.rs2) as i32) {
+                cpu.add2pc(inst.imm.unwrap() as u32);
             } 
         },
         OpecodeKind::OP_BLTU => {
-            if (cpu.regs.read(inst.rs1) as u32) < (cpu.regs.read(inst.rs2) as u32) {
-                cpu.add2pc(inst.imm.unwrap());
+            if cpu.regs.read(inst.rs1) < cpu.regs.read(inst.rs2) {
+                cpu.add2pc(inst.imm.unwrap() as u32);
             } 
         },
         OpecodeKind::OP_BGEU => {
-            if (cpu.regs.read(inst.rs1) as u32) >= (cpu.regs.read(inst.rs2) as u32) {
-                cpu.add2pc(inst.imm.unwrap());
+            if cpu.regs.read(inst.rs1) >= cpu.regs.read(inst.rs2) {
+                cpu.add2pc(inst.imm.unwrap() as u32);
             } 
         },
         OpecodeKind::OP_LB => {
-            let load_addr = cpu.trans_addr(TransFor::Load, cpu.regs.read(inst.rs1) + inst.imm.unwrap())?;
+            let load_addr = cpu.trans_addr(TransFor::Load, (cpu.regs.read(inst.rs1) as i32 + inst.imm.unwrap()) as u32)?;
             cpu.regs.write(inst.rd, cpu.bus.load8(load_addr)?);
         },
         OpecodeKind::OP_LH => {
-            let load_addr = cpu.trans_addr(TransFor::Load, cpu.regs.read(inst.rs1) + inst.imm.unwrap())?;
+            let load_addr = cpu.trans_addr(TransFor::Load, (cpu.regs.read(inst.rs1) as i32 + inst.imm.unwrap()) as u32)?;
             cpu.regs.write(inst.rd, cpu.bus.load16(load_addr)?);
         },
         OpecodeKind::OP_LW => {
-            let load_addr = cpu.trans_addr(TransFor::Load, cpu.regs.read(inst.rs1) + inst.imm.unwrap())?;
+            let load_addr = cpu.trans_addr(TransFor::Load, (cpu.regs.read(inst.rs1) as i32 + inst.imm.unwrap()) as u32)?;
             cpu.regs.write(inst.rd, cpu.bus.load32(load_addr)?);
         },
         OpecodeKind::OP_LBU => {
-            let load_addr = cpu.trans_addr(TransFor::Load, cpu.regs.read(inst.rs1) + inst.imm.unwrap())?;
+            let load_addr = cpu.trans_addr(TransFor::Load, cpu.regs.read(inst.rs1) + inst.imm.unwrap() as u32)?;
             cpu.regs.write(inst.rd, cpu.bus.load_u8(load_addr)?);
         },
         OpecodeKind::OP_LHU => {
-            let load_addr = cpu.trans_addr(TransFor::Load, cpu.regs.read(inst.rs1) + inst.imm.unwrap())?;
+            let load_addr = cpu.trans_addr(TransFor::Load, cpu.regs.read(inst.rs1) + inst.imm.unwrap() as u32)?;
             cpu.regs.write(inst.rd, cpu.bus.load_u16(load_addr)?);
         },
         OpecodeKind::OP_SB => {
-            let store_addr = cpu.trans_addr(TransFor::StoreAMO, cpu.regs.read(inst.rs1) + inst.imm.unwrap())?;
+            let store_addr = cpu.trans_addr(TransFor::StoreAMO, (cpu.regs.read(inst.rs1) as i32 + inst.imm.unwrap()) as u32)?;
             cpu.bus.store8(store_addr, cpu.regs.read(inst.rs2))?;
         },
         OpecodeKind::OP_SH => {
-            let store_addr = cpu.trans_addr(TransFor::StoreAMO, cpu.regs.read(inst.rs1) + inst.imm.unwrap())?;
+            let store_addr = cpu.trans_addr(TransFor::StoreAMO, (cpu.regs.read(inst.rs1) as i32 + inst.imm.unwrap()) as u32)?;
             cpu.bus.store16(store_addr, cpu.regs.read(inst.rs2))?;
         },
         OpecodeKind::OP_SW => {
-            let store_addr = cpu.trans_addr(TransFor::StoreAMO, cpu.regs.read(inst.rs1) + inst.imm.unwrap())?;
+            let store_addr = cpu.trans_addr(TransFor::StoreAMO, (cpu.regs.read(inst.rs1) as i32 + inst.imm.unwrap()) as u32)?;
             cpu.bus.store32(store_addr, cpu.regs.read(inst.rs2))?;
         },
         OpecodeKind::OP_ADDI => {
-            cpu.regs.write(inst.rd, cpu.regs.read(inst.rs1) + inst.imm.unwrap());
+            cpu.regs.write(inst.rd, (cpu.regs.read(inst.rs1) as i32 + inst.imm.unwrap()) as u32);
         },
         OpecodeKind::OP_SLTI => {
             cpu.regs.write(inst.rd,  
-                (cpu.regs.read(inst.rs1) < inst.imm.unwrap()) as i32);
+                ((cpu.regs.read(inst.rs1) as i32) < (inst.imm.unwrap() as i32)) as u32);
         },
         OpecodeKind::OP_SLTIU => {
             cpu.regs.write(inst.rd,  
-                ((cpu.regs.read(inst.rs1) as u32) < inst.imm.unwrap() as u32) as i32);
+                (cpu.regs.read(inst.rs1) < (inst.imm.unwrap() as u32)) as u32);
         },
         OpecodeKind::OP_XORI => {
-            cpu.regs.write(inst.rd, cpu.regs.read(inst.rs1) ^ inst.imm.unwrap());
+            cpu.regs.write(inst.rd, cpu.regs.read(inst.rs1) ^ inst.imm.unwrap() as u32);
         },
         OpecodeKind::OP_ORI => {
-            cpu.regs.write(inst.rd, cpu.regs.read(inst.rs1) | inst.imm.unwrap());
+            cpu.regs.write(inst.rd, cpu.regs.read(inst.rs1) | inst.imm.unwrap() as u32);
         },
         OpecodeKind::OP_ANDI => {
-            cpu.regs.write(inst.rd, cpu.regs.read(inst.rs1) & inst.imm.unwrap());
+            cpu.regs.write(inst.rd, cpu.regs.read(inst.rs1) & inst.imm.unwrap() as u32);
         },
         OpecodeKind::OP_SLLI => {
             cpu.regs.write(inst.rd,
-                ((cpu.regs.read(inst.rs1) as u32) << inst.imm.unwrap()) as i32);
-        },                                                
+                (cpu.regs.read(inst.rs1) << inst.imm.unwrap() as u32));
+        },                                        
         OpecodeKind::OP_SRLI => {
             cpu.regs.write(inst.rd,
-                ((cpu.regs.read(inst.rs1) as u32) >> inst.imm.unwrap()) as i32);
+                (cpu.regs.read(inst.rs1) >> inst.imm.unwrap() as u32));
         },
         OpecodeKind::OP_SRAI => {
             cpu.regs.write(inst.rd,
-                ((cpu.regs.read(inst.rs1) as i32) >> inst.imm.unwrap()) as i32);
+                ((cpu.regs.read(inst.rs1) as i32) >> inst.imm.unwrap()) as u32);
         },
         OpecodeKind::OP_ADD => {
             cpu.regs.write(inst.rd,
@@ -126,15 +126,15 @@ pub fn exec(inst: &Instruction, cpu: &mut CPU) -> Result<(), (Option<u32>, TrapC
         },
         OpecodeKind::OP_SLL => {
             cpu.regs.write(inst.rd,
-                ((cpu.regs.read(inst.rs1) as u32) << cpu.regs.read(inst.rs2)) as i32);
+                (cpu.regs.read(inst.rs1) << cpu.regs.read(inst.rs2)));
         },
         OpecodeKind::OP_SLT => {
             cpu.regs.write(inst.rd,
-                (cpu.regs.read(inst.rs1) < cpu.regs.read(inst.rs2)) as i32);
+                ((cpu.regs.read(inst.rs1) as i32) < (cpu.regs.read(inst.rs2) as i32)) as u32);
         },
         OpecodeKind::OP_SLTU => {
             cpu.regs.write(inst.rd,
-                ((cpu.regs.read(inst.rs1) as u32) < (cpu.regs.read(inst.rs2) as u32)) as i32);
+                (cpu.regs.read(inst.rs1) < cpu.regs.read(inst.rs2)) as u32);
         },
         OpecodeKind::OP_XOR => {
             cpu.regs.write(inst.rd,
@@ -142,11 +142,11 @@ pub fn exec(inst: &Instruction, cpu: &mut CPU) -> Result<(), (Option<u32>, TrapC
         },
         OpecodeKind::OP_SRL => {
             cpu.regs.write(inst.rd,
-                ((cpu.regs.read(inst.rs1) as u32) >> cpu.regs.read(inst.rs2)) as i32);
+                (cpu.regs.read(inst.rs1) >> cpu.regs.read(inst.rs2)));
         },
         OpecodeKind::OP_SRA => {
             cpu.regs.write(inst.rd,
-                (cpu.regs.read(inst.rs1) as i32) >> cpu.regs.read(inst.rs2));
+                ((cpu.regs.read(inst.rs1) as i32) >> cpu.regs.read(inst.rs2)) as u32);
         },
         OpecodeKind::OP_OR => {
             cpu.regs.write(inst.rd,
